moved the apps under cmd and the provider packages under pkg
